refact: move shared/ui/styles/utils to shared/lib/cn
