+++ src/features/category/ui/CategoryTag/CategoryTag.tsx
@@ -1,7 +1,7 @@
 import { Badge } from '@/shared/ui/atoms/badge';
 import { Button } from '@/shared/ui/atoms/button';
 import X from '@/shared/ui/icons/X';
-import { cn } from '@/shared/ui/styles/utils';
+import { cn } from '@/shared/lib/cn';
 
 export interface CategoryTagProps {
   tagName: string;
