+++ src/features/expense/ui/ConsumptionKindDrawer/ConsumptionKindDrawer.tsx
@@ -28,7 +28,7 @@ import {
 import Check from '@/shared/ui/icons/Check';
 import Exclamation from '@/shared/ui/icons/Exclamation';
 import X from '@/shared/ui/icons/X';
-import { cn } from '@/shared/ui/styles/utils';
+import { cn } from '@/shared/lib/cn';
 
 interface RadioItemProps {
   isSelected: boolean;
