+++ src/features/expense/ui/ExpenseForm/ExpenseForm.tsx
@@ -22,7 +22,7 @@ import {
 } from '@/shared/ui/atoms/form';
 import ArrowRight from '@/shared/ui/icons/ArrowRight';
 import LabeledTextarea from '@/shared/ui/LabeledTextarea';
-import { cn } from '@/shared/ui/styles/utils';
+import { cn } from '@/shared/lib/cn';
 
 const CalendarDrawerTrigger = ({ date }: { date: Date }) => {
   return (
