+++ src/features/onboarding/ui/Onboarding.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useRef, useState } from 'react';
 
-import { cn } from '@/shared/ui/styles/utils';
+import { cn } from '@/shared/lib/cn';
 
 interface OnboardingProps {
   onComplete?: () => void;
