+++ src/shared/ui/atoms/calendar.tsx
@@ -4,7 +4,7 @@ import { DayPicker } from 'react-day-picker';
 import { buttonVariants } from '@/shared/ui/atoms/button';
 import ArrowLeft from '@/shared/ui/icons/ArrowLeft';
 import ArrowRight from '@/shared/ui/icons/ArrowRight';
-import { cn } from '@/shared/ui/styles/utils';
+import { cn } from '@/shared/lib/cn';
 
 function Calendar({
   className,
