+++ src/shared/ui/atoms/input-tags.tsx
@@ -8,7 +8,7 @@ import * as React from 'react';
 
 import { Badge } from '@/shared/ui/atoms/badge';
 import { Button } from '@/shared/ui/atoms/button';
-import { cn } from '@/shared/ui/styles/utils';
+import { cn } from '@/shared/lib/cn';
 
 import { type InputProps } from './input';
