+++ src/shared/ui/atoms/input.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { cn } from '@/shared/ui/styles/utils';
+import { cn } from '@/shared/lib/cn';
 
 export type InputProps = React.ComponentProps<'input'>;
