+++ src/shared/ui/atoms/label.tsx
@@ -1,7 +1,7 @@
 import * as LabelPrimitive from '@radix-ui/react-label';
 import * as React from 'react';
 
-import { cn } from '@/shared/ui/styles/utils';
+import { cn } from '@/shared/lib/cn';
 
 function Label({
   className,
