+++ src/shared/ui/atoms/textarea.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { cn } from '@/shared/ui/styles/utils';
+import { cn } from '@/shared/lib/cn';
 
 function Textarea({ className, ...props }: React.ComponentProps<'textarea'>) {
   return (
